Update language detection in API
Previous versions of retext-language detected languages on
paragraphs and documents by calculating the average of their
children (sentences, paragraphs). This wasn't working for
multi-language parents. For example, when a paragraph included
two sentences, one in English and one in French, their parent
would get their average, something like Catalan.

Now, the API only looks at the primary language for a sentence,
which in my tests results in better results. The paragraph's
primary language is either French or English, and it's languages
are both French and English.